Split scanning and stats into their own files
