+++ scripts/get-usage/types.ts
@@ -1,0 +1,1 @@
+export type ComponentUsageData = Record<string, Record<string, number>>;
